TP du 23/06/2020

+++ src/fr/declaration/variable/DeclarationApp.java
@@ -1,19 +1,19 @@
 package fr.declaration.variable;
 
+import fr.boucles.ExerciceBoucleBase;
 public class DeclarationApp {
 	public static void main(String[] args) {
 		byte var1 = 12;
 		short var2 = 20000;
 		int var3 = 100000;
-		long var4 = 1000000;
-		float var5 = 2.5f;
-		double var6 = 2.55d;
+		long var4 = 1000000l;
+		double var5 = 2.5;
+		float var6 = 2.55f;
 		char var7 = 'f';
 		boolean var8 = true;
 		String var9 = "hello";
 		
 		System.out.println(var6);
 		System.out.println("Salut à tous");
-		
 	}
 }
